chore: clean numeric variable types

+++ api/models/menu_item.py
@@ -8,7 +8,7 @@ class MenuItem(Base):
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     name = Column(String(100), nullable=False)
-    price = Column(Integer, nullable=False)
+    price = Column(DECIMAL(10, 2), nullable=False)
     category = Column(String(100), nullable=False)
     calories = Column(Integer)
